change code

+++ quanlyphonggym/src/main/java/quanlyphonggym/Controllers/AdminCtrl/QuanLyGoiTap/CRUDGoiTapCtrl.java
@@ -6,6 +6,7 @@ import quanlyphonggym.MysqlConnection;
 import javax.swing.*;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public class CRUDGoiTapCtrl {
@@ -58,10 +59,29 @@ public class CRUDGoiTapCtrl {
 
     public boolean deleteGoiTap(int idGoiTap) throws SQLException, ClassNotFoundException {
         Connection connection = MysqlConnection.getMysqlConnection();
-        String sql = "DELETE FROM goitap WHERE id = " +idGoiTap;
-        PreparedStatement preparedStatement = connection.prepareStatement(sql);
-        preparedStatement.execute();
-        preparedStatement.close();
+
+        //check đã có học viên nào đăng kí gói chưa
+        try {
+            String sqlCheckHV = "SELECT * FROM dangkygoitap WHERE idGoiTap = " + idGoiTap;
+            PreparedStatement prds = connection.prepareStatement(sqlCheckHV);
+            ResultSet ertr = prds.executeQuery();
+            if (ertr.next()) {
+                JOptionPane.showMessageDialog(null, "Đã có học viên đăng kí gói tập này. Hãy hủy các đăng kí trước khi xóa");
+                prds.close();
+                return false;
+            }
+
+            String sql = "DELETE FROM goitap WHERE id = " + idGoiTap;
+            PreparedStatement preparedStatement = connection.prepareStatement(sql);
+            preparedStatement.execute();
+            preparedStatement.close();
+        } catch (SQLException e) {
+            JOptionPane.showMessageDialog(null, "Đã có lỗi xảy ra");
+            System.out.println(e.getMessage());
+            connection.close();
+            return false;
+        }
+        connection.close();
         return true;
     }
 }
